refactor(pipeline): move plugins to plugin/ directory
