+++ packages/pipeline/src/index.ts
@@ -9,5 +9,5 @@ export * from './stageOutputResolver.js';
 export * from './sparql/index.js';
 export * from './distribution/index.js';
 export * from './writer/index.js';
-export * from './provenance.js';
-export * from './schemaOrgNormalization.js';
+export * from './plugin/provenance.js';
+export * from './plugin/schemaOrgNormalization.js';
