+++ packages/pipeline/src/plugin/schemaOrgNormalization.ts
@@ -1,5 +1,5 @@
-import type { QuadTransform } from './stage.js';
-import type { PipelinePlugin } from './pipeline.js';
+import type { QuadTransform } from '../stage.js';
+import type { PipelinePlugin } from '../pipeline.js';
 import type { Quad } from '@rdfjs/types';
 import { DataFactory } from 'n3';
